before files rearenge
